+++ packages/component/src/hooks/index.js
@@ -9,6 +9,8 @@ import useLastTypingAt from './useLastTypingAt';
 import useLocalize from './useLocalize';
 import useLocalizeDate from './useLocalizeDate';
 import usePerformCardAction from './usePerformCardAction';
+import useConnectivityStatus from './useConnectivityStatus';
+import useGroupTimestamp from './useGroupTimestamp';
 import usePostActivity from './usePostActivity';
 import useReferenceGrammarID from './useReferenceGrammarID';
 import useRenderMarkdownAsHTML from './useRenderMarkdownAsHTML';
@@ -24,6 +26,9 @@ import useStyleOptions from './useStyleOptions';
 import useStyleSet from './useStyleSet';
 import useSubmitSendBox from './useSubmitSendBox';
 import useSuggestedActions from './useSuggestedActions';
+import useTimeoutForSend from './useTimeoutForSend';
+import useUserID from './useUserID';
+import useUsername from './useUsername';
 
 import { useSendBoxDictationStarted } from '../BasicSendBox';
 import { useTextBoxSubmit } from '../SendBox/TextBox';
@@ -36,6 +41,8 @@ export {
   useDisabled,
   useEmitTypingIndicator,
   useFocusSendBox,
+  useConnectivityStatus,
+  useGroupTimestamp,
   useLanguage,
   useLastTypingAt,
   useLocalize,
@@ -58,5 +65,8 @@ export {
   useSubmitSendBox,
   useSuggestedActions,
   useTextBoxSubmit,
-  useTypingIndicatorVisible
+  useTimeoutForSend,
+  useTypingIndicatorVisible,
+  useUserID,
+  useUsername
 };